Remove dead code from ota.c
This NULL check is unreachable and not required. This check can
never be triggered because the if condition that comes before it
is dereferencing the pointer that it is about to check. If it was
NULL, there would have been an issue already.

The check is not required because the pointer is to the address of
a static variable that is allocated in the global scope.