posix: Cast timer_t to (void *) before assigning to sigevent.sigev_value.sival_ptr
sigevent.sigev_value.sival_ptr is of type (void *) per POSIX spec.
Currently the codes assign otaTimers[ otaTimerId ] (type timer_t) to
it without any casting.

The POSIX spec does not clearly define what type timer_t should be so
it seems to be an implementation specific type. On platforms like Linux
with glibc the timer_t is (void *) but on some other platforms this
might be an integer, like in VxWorks it is unsigned long. On such
platforms it unfortunately creates a warning during build:

  warning: incompatible integer to pointer conversion assigning to
  'void *' from 'timer_t' (aka 'unsigned long') [-Wint-conversion]
      sgEvent.sigev_value.sival_ptr = otaTimers[ otaTimerId ];

Casting timer_t to (void *) before the assignment.

Signed-off-by: Yiting Wang <yiting.wang@windriver.com>